+++ src/catalog/rest/catalog_entry/table/iceberg_table_entry.cpp
@@ -21,6 +21,7 @@
 #include "catalog/rest/storage/authorization/sigv4.hpp"
 #include "rest_catalog/objects/list.hpp"
 #include "catalog/rest/catalog_entry/table/iceberg_table_information.hpp"
+#include "catalog/rest/storage/iceberg_table_secret_provider.hpp"
 #include "catalog/rest/transaction/iceberg_transaction.hpp"
 
 namespace duckdb {
@@ -35,16 +36,6 @@ IcebergTableEntry::IcebergTableEntry(IcebergTableInformation &table_info, Catalo
 
 unique_ptr<BaseStatistics> IcebergTableEntry::GetStatistics(ClientContext &context, column_t column_id) {
 	return nullptr;
-}
-
-void AddHTTPSecretsToOptions(SecretEntry &http_secret_entry, case_insensitive_map_t<Value> &options) {
-	auto http_kv_secret = dynamic_cast<const KeyValueSecret &>(*http_secret_entry.secret);
-
-	options["http_proxy"] =
-	    http_kv_secret.TryGetValue("http_proxy").IsNull() ? "" : http_kv_secret.TryGetValue("http_proxy").ToString();
-	options["verify_ssl"] = http_kv_secret.TryGetValue("verify_ssl").IsNull()
-	                            ? Value::BOOLEAN(true)
-	                            : http_kv_secret.TryGetValue("verify_ssl").DefaultCastAs(LogicalType::BOOLEAN);
 }
 
 void IcebergTableEntry::PrepareIcebergScanFromEntry(ClientContext &context) const {
@@ -142,7 +133,7 @@ void IcebergTableEntry::PrepareIcebergScanFromEntry(ClientContext &context) cons
 		}
 
 		if (http_secret_entry) {
-			AddHTTPSecretsToOptions(*http_secret_entry, info.options);
+			IcebergTableSecretProvider::AddHTTPSecretsToOptions(*http_secret_entry, info.options);
 		}
 
 		auto created_secret = secret_manager.CreateSecret(context, info);
@@ -162,7 +153,7 @@ void IcebergTableEntry::PrepareIcebergScanFromEntry(ClientContext &context) cons
 	} else {
 		for (auto &info : table_credentials.storage_credentials) {
 			if (http_secret_entry) {
-				AddHTTPSecretsToOptions(*http_secret_entry, info.options);
+				IcebergTableSecretProvider::AddHTTPSecretsToOptions(*http_secret_entry, info.options);
 			}
 			auto created_secret = secret_manager.CreateSecret(context, info);
 			transaction.created_secrets.insert(created_secret->secret->GetName());
