+++ test/sql/local/catalog_custom_setup/fixture/vended_credentials_refresh.test
@@ -29,7 +29,10 @@ CALL enable_logging('HTTP')
 statement ok
 CREATE SECRET vended_refresh_http_proxy (
     TYPE HTTP,
-    HTTP_PROXY '${VENDED_CREDENTIAL_REFRESH_PROXY}'
+    HTTP_PROXY '${VENDED_CREDENTIAL_REFRESH_PROXY}',
+    HTTP_PROXY_USERNAME 'refresh_proxy_user',
+    HTTP_PROXY_PASSWORD 'refresh_proxy_password',
+    EXTRA_HTTP_HEADERS MAP {'x-refresh-test': 'present'}
 )
 
 statement ok
@@ -97,15 +100,19 @@ FROM duckdb_logs_parsed('HTTP')
 WHERE starts_with(request.url, 'http://127.0.0.1:9000/')
   AND response.status IN ('OK_200', 'PartialContent_206')
   AND request.headers['Authorization'] LIKE '%NEW_INIT_KEY%'
+  AND request.headers['x-refresh-test'] = 'present'
 ----
 1
 
-query II
-SELECT count(*), CAST(bool_or(secret_string LIKE '%key_id=NEW_INIT_KEY%') AS INTEGER)
+query IIII
+SELECT count(*),
+       CAST(bool_or(secret_string LIKE '%key_id=NEW_INIT_KEY%') AS INTEGER),
+       CAST(bool_or(secret_string LIKE '%http_proxy_username=refresh_proxy_user%') AS INTEGER),
+       CAST(bool_or(secret_string LIKE '%x-refresh-test%') AS INTEGER)
 FROM duckdb_secrets(redact=false)
 WHERE name='vended_init_refresh_secret'
 ----
-1	1
+1	1	1	1
 
 query I
 SELECT count(*)
@@ -178,15 +185,19 @@ WHERE starts_with(request.url, 'http://127.0.0.1:9000/')
   AND request.headers['Range'] IS NOT NULL
   AND response.status = 'PartialContent_206'
   AND request.headers['Authorization'] LIKE '%NEW_RANGE_KEY%'
+  AND request.headers['x-refresh-test'] = 'present'
 ----
 1
 
-query II
-SELECT count(*), CAST(bool_or(secret_string LIKE '%key_id=NEW_RANGE_KEY%') AS INTEGER)
+query IIII
+SELECT count(*),
+       CAST(bool_or(secret_string LIKE '%key_id=NEW_RANGE_KEY%') AS INTEGER),
+       CAST(bool_or(secret_string LIKE '%http_proxy_username=refresh_proxy_user%') AS INTEGER),
+       CAST(bool_or(secret_string LIKE '%x-refresh-test%') AS INTEGER)
 FROM duckdb_secrets(redact=false)
 WHERE name='vended_range_refresh_secret'
 ----
-1	1
+1	1	1	1
 
 query I
 SELECT count(*)
